Put expenses functions in different file
