+++ src/past-trips.js
@@ -30,56 +30,6 @@ function viewPreviousTrip(id, allTrips, allDestinations){
     };
 };
 
-function calculateAnnualTripCost(id, allTrips, allDestinations) {
-    let allPastTrips = allTrips.filter(trip => {
-        return trip.userID === id && trip.status === "approved"; 
-    }).sort((a, b) => new Date(b.date) - new Date(a.date))
-    let year = allPastTrips[0].date.split('/')[0]
-    let annualTrips = allPastTrips.filter(trip => {
-        return trip.date.split('/')[0] === year; 
-    });
-    let totalTripCost = annualTrips.reduce((total, trip) => {
-        return total += ((trip.duration * (allDestinations.find(destination => {
-        return destination.id === trip.destinationID
-       })).estimatedLodgingCostPerDay) + allDestinations.find(destination =>{
-        return destination.id === trip.destinationID
-       }).estimatedFlightCostPerPerson) 
-    }, 0);
-    return (totalTripCost + (totalTripCost * .10)).toFixed(2);
-};
-
-function calculateAnnualFlightCost(id, allTrips, allDestinations){
-    let allPastTrips = allTrips.filter(trip => {
-        return trip.userID === id && trip.status === "approved"; 
-    }).sort((a, b) => new Date(b.date) - new Date(a.date))
-    let year = allPastTrips[0].date.split('/')[0]
-    let annualTrips = allPastTrips.filter(trip => {
-        return trip.date.split('/')[0] === year; 
-    });
-    let totalFlightCost = annualTrips.reduce((total, trip) => {
-        return total += allDestinations.find(destination => {
-        return destination.id === trip.destinationID
-       }).estimatedFlightCostPerPerson
-    }, 0);
-    return totalFlightCost.toFixed(2)
-}
-
-function calculateAnnualLodgingCost(id, allTrips, allDestinations){
-    let allPastTrips = allTrips.filter(trip => {
-        return trip.userID === id && trip.status === "approved"; 
-    }).sort((a, b) => new Date(b.date) - new Date(a.date))
-    let year = allPastTrips[0].date.split('/')[0]
-    let annualTrips = allPastTrips.filter(trip => {
-        return trip.date.split('/')[0] === year; 
-    });
-    let totalLodgingCost = annualTrips.reduce((total, trip) => {
-        return total += trip.duration * (allDestinations.find(destination => {
-        return destination.id === trip.destinationID
-       })).estimatedLodgingCostPerDay
-    }, 0);
-    return totalLodgingCost.toFixed(2);
-}
-
 function viewUpcomingTrip(id, allTrips){
     let allPastTrips = allTrips.filter(trip => {
         return trip.userID === id && trip.status === "approved"; 
@@ -94,5 +44,5 @@ function viewUpcomingTrip(id, allTrips){
     return pastTripDestination.slice(-1);
 };
 
-export { viewPastTrips, viewUpcomingTrip, calculateAnnualTripCost, viewPreviousTrip, calculateAnnualFlightCost, calculateAnnualLodgingCost }
+export { viewPastTrips, viewUpcomingTrip, viewPreviousTrip }
