+++ test/trip-history.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 const expect = chai.expect;
-import { viewPastTrips, viewUpcomingTrip, calculateAnnualTripCost, viewPreviousTrip, calculateAnnualFlightCost, calculateAnnualLodgingCost } from '../src/past-trips';
+import { viewPastTrips, viewUpcomingTrip, viewPreviousTrip } from '../src/past-trips';
 import { trips } from '../src/data/sample-trips';
 import { destinations } from '../src/data/sample-destinations';
 
@@ -53,34 +53,6 @@ describe('trip-history.js', function() {
 
       expect(noPreviousTrip).to.equal(`You have not been on any trips yet.`)
     })
-  });
-
-  describe('calculate annual trip cost', function() {
-    it('should calculate the cost of travel for the year for a single traveler', function() {
-      const totalCost = calculateAnnualTripCost(6, trips, destinations)
-      const totalCost2 = calculateAnnualTripCost(5, trips, destinations)
-
-      expect(totalCost).to.equal("2546.50")
-      expect(totalCost2).to.equal("6385.50")
-    });
-  });
-
-  describe('calculate annual flight cost', function() {
-    it('should calculate the cost of flights for the year for a single traveler', function() {
-      const flightCost = calculateAnnualFlightCost(2, trips, destinations)
-
-      expect(flightCost).to.equal("1750.00")
-    });
-  });
-
-  describe('calculate annual loding cost', function() {
-    it('should calculate the cost of lodging for the year for a single traveler', function() {
-      const lodgingCost = calculateAnnualLodgingCost(1, trips, destinations);
-      const lodgingCost2 = calculateAnnualLodgingCost(2, trips, destinations);
-
-      expect(lodgingCost).to.equal("800.00")
-      expect(lodgingCost2).to.equal("2550.00")
-    });
   });
 
   describe('find upcoming trip', function() {
